+++ surepy/tests/analysis/test_Ripley.py
@@ -6,7 +6,7 @@ from surepy import LocData
 import surepy.constants
 import surepy.io.io_locdata as io
 import surepy.tests.test_data
-from surepy.analysis import Ripleys_k_function, Ripleys_l_function, Ripleys_h_function
+from surepy.analysis import RipleysKFunction, RipleysLFunction, RipleysHFunction
 
 
 @pytest.fixture()
@@ -20,32 +20,32 @@ def locdata_blobs_3D():
 
 def test_Ripleys_k_function(locdata_blobs):
     radii = np.linspace(0,100,20)
-    rhf = Ripleys_k_function(locdata_blobs, radii=radii).compute()
+    rhf = RipleysKFunction(locdata_blobs, radii=radii).compute()
     #print(rhf.results[0:5])
     assert (len(rhf.results) == len(radii))
     #rhf.plot()
 
 def test_Ripleys_k_function_3d(locdata_blobs_3D):
     radii = np.linspace(0,100,20)
-    rhf = Ripleys_k_function(locdata_blobs_3D, radii=radii).compute()
+    rhf = RipleysKFunction(locdata_blobs_3D, radii=radii).compute()
     #print(rhf.results[0:5])
     assert (len(rhf.results) == len(radii))
 
 def test_Ripleys_k_function_estimate(locdata_blobs):
     radii = np.linspace(0,100,20)
-    rhf = Ripleys_k_function(locdata_blobs, radii=radii, region_measure=1, other_locdata=locdata_blobs).compute()
+    rhf = RipleysKFunction(locdata_blobs, radii=radii, region_measure=1, other_locdata=locdata_blobs).compute()
     #print(rhf.results[0:5])
     assert (len(rhf.results) == len(radii))
 
 def test_Ripleys_l_function(locdata_blobs):
     radii = np.linspace(0,100,20)
-    rhf = Ripleys_l_function(locdata_blobs, radii=radii).compute()
+    rhf = RipleysLFunction(locdata_blobs, radii=radii).compute()
     #print(rhf.results[0:5])
     assert (len(rhf.results) == len(radii))
 
 def test_Ripleys_h_function(locdata_blobs):
     radii = np.linspace(0,100,20)
-    rhf = Ripleys_h_function(locdata_blobs, radii=radii).compute()
+    rhf = RipleysHFunction(locdata_blobs, radii=radii).compute()
     #print(rhf._Ripley_h_maximum)
     #print(rhf.Ripley_h_maximum)
     #print(rhf.results[5:11])